Fixes a small timing bug during shutdown.
There was a small race condition in the previous implementation. This
race condition did *not* result in bugs given the current shutdown
algorithm. However, it could lead to subtle bugs should that algorithm
change. This PR imposes a small performance penalty on every job post to
a queue but it is the most "correct" approach.