Setup alternative shutdown strategy #203
Currently, when the CountDownLatch job is posted to each queue, if the
queue has more than 1 worker, it'll get picked up by a worker in
parallel to one that may already be in progress and long-running. This
has the potential to cut off the last work in each queue upon shutdown
depending on the speed of that job.

This strategy leans on the timeout while checking if each queue is idle
to decide whether to shutdown.